Add files via upload
This script is a powerful tool that can help you increase engagement and views for your Instagram page in a real and free way. It works by accessing the data of your Instagram page and generating a direct link to your latest posts, which is then automatically copied to your clipboard. This enables you to discover more people and increase their engagement with your posts.

The script is easy to use and doesn't require you to log in to your Instagram account. It uses modern techniques based on artificial intelligence and machine learning to optimize the appearance of your posts in the Instagram Explore page in a smart and effective way.

By using this script, you can increase engagement and views for your page in a real and free way, making it a powerful and effective tool to help grow your Instagram page.